python-service/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pdfminer.six
spacy
sentence-transformers
torch